Refactored run command

+++ src/command/run/operator/save.py
@@ -5,11 +5,11 @@ from enum import Enum
 from typing import Final
 
 import duckdb
-from pydantic import BaseModel
 
 from settings import settings
 from src import robolog
 from src.command.run import validate
+from src.command.run.operator.base import Operator
 
 
 class FileExtension(Enum):
@@ -20,7 +20,7 @@ class FileExtension(Enum):
     JSONL = "jsonl"
 
 
-class SaveDataFrame(BaseModel):
+class SaveDataFrame(Operator):
     """An operator that saves a DuckDB view to disk."""
 
     # Name of the DuckDB view to save
@@ -35,16 +35,16 @@ class SaveDataFrame(BaseModel):
     @property
     def running_status(self) -> str:
         """Return the running status of the operator."""
-        return f"Saving [bold]{self.name}[/bold] to disk"
+        return f"Saving **{self.name}** to disk"
 
     @property
     def finished_status(self) -> str:
         """Return the finished status of the operator."""
-        return f":floppy_disk: Saved [bold]{self.name}[/bold] to {self._saved_file_path}"
+        return f":floppy_disk: Saved **{self.name}** to {self._saved_file_path}"
 
     @staticmethod
-    def from_dict(data: str) -> "SaveDataFrame":
-        """Return a SaveDataFrame instance from a dictionary."""
+    def from_name(data: str) -> "SaveDataFrame":
+        """Return a SaveDataFrame instance from a DataFrame name."""
         validate.validate_snake_case(data)
         return SaveDataFrame(name=data)
 
@@ -54,30 +54,33 @@ class SaveDataFrame(BaseModel):
         start_seconds: float | None,
         end_seconds: float | None,
         ext: FileExtension = FileExtension.PARQUET,
+        dry_run: bool = False,
     ) -> pathlib.Path:
         """Write the DuckDB view to disk and return the file path."""
         file_path = (
-            pathlib.Path(settings.STORAGE_DIRECTORY)
+            pathlib.Path(settings.DATASET_DIRECTORY)
             / self.name
             / f"datestr={robolog.datestr(robolog_path)}"
             / f"{robolog.snippet_name(robolog_path, start_seconds, end_seconds)}.{ext.value}"
         )
 
         try:
-            file_path.parent.mkdir(parents=True, exist_ok=True)
             relation = duckdb.view(self.name)
-            match ext:
-                case FileExtension.PARQUET:
-                    relation.write_parquet(str(file_path))
-                case FileExtension.CSV:
-                    relation.write_csv(str(file_path))
-                case FileExtension.JSONL:
-                    duckdb.sql(f"COPY relation TO '{file_path!s}' (FORMAT 'JSON')")
-                case _:
-                    raise ValueError(f"Unsupported disk format: {ext}")
+            if not dry_run:
+                file_path.parent.mkdir(parents=True, exist_ok=True)
+                match ext:
+                    case FileExtension.PARQUET:
+                        relation.write_parquet(str(file_path))
+                    case FileExtension.CSV:
+                        relation.write_csv(str(file_path))
+                    case FileExtension.JSONL:
+                        duckdb.sql(f"COPY relation TO '{file_path!s}' (FORMAT 'JSON')")
+                    case _:
+                        raise ValueError(f"Unsupported disk format: {ext}")
             self._saved_file_path = file_path
 
         except Exception as e:
             self._saved_file_path = None
-            file_path.unlink(missing_ok=True)
+            if not dry_run:
+                file_path.unlink(missing_ok=True)
             raise e
